--- unit-test/data/trackerAllowlist.json
@@ -1,26 +1,0 @@
-{
-    "allowlist-tracker-1.com": {
-        "rules" : [
-            {
-                "rule": "allowlist-tracker-1.com/videos.js",
-                "domains": ["testsite.com"],
-                "reason": "match single resource on single site"
-            }
-        ]
-    },
-    "allowlist-tracker-2.com": {
-        "rules" : [
-            {
-                "rule": "allowlist-tracker-2.com/comments/",
-                "domains": ["<all>"],
-                "reason": "match all sites and all paths"
-            },
-            {
-                "rule": "allowlist-tracker-2.com/login.js",
-                "domains": ["<all>"],
-                "reason": "match single resource on all sites"
-            }
-        ]
-    }
-}
-
